Revert "feat: бновление библиотеки audioI2S до версии 3.3.2r (Jul 18.2025)"
This reverts commit 21643cc2ea04f01cc274a3c2524bc68ff3032a26.

+++ src/src/displays/tools/spectrum_analyzer.h
@@ -67,7 +67,7 @@ public:
     bool begin();
     
     // Обработка аудио
-    void processAudio(const int16_t* samples, int32_t count);
+    void processAudio(const int16_t* samples, uint16_t count);
     
     // Сброс данных спектра
     void clearData();
